greent/requirements.txt
appnope==0.1.0
autovizwidget==0.12.4
bleach==2.1.1
bravado==9.0.2
bravado-core==4.9.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.1.2
entrypoints==0.2.3
Flask==0.12
Flask-GraphQL==1.4.1
Flask-Testing==0.6.2
gnureadline==6.3.3
graphene==1.4.1
graphql-core==1.1
graphql-relay==0.4.5
green==2.11.0
hdijupyterutils==0.12.4
html5lib==1.0b10
idna==2.5
ipykernel==4.6.1
ipython==4.2.1
ipython-genutils==0.2.0
ipywidgets==6.0.1
isodate==0.5.4
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.5.1
MarkupSafe==1.0
mistune==0.7.4
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==4.4.1
numpy==1.13.3
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
pkginfo==1.4.1
plotly==1.10.0
promise==2.0.2
prompt-toolkit==1.0.15
pronto==0.8.0
ptyprocess==0.5.2
Pygments==2.2.0
pykerberos==1.1.14
PyLD==0.7.3
pyparsing==2.2.0
python-dateutil==2.6.1
python-termstyle==0.1.10
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
rdflib==4.2.2
requests==2.18.1
requests-kerberos==0.11.0
requests-toolbelt==0.8.0
rfc3987==1.3.7
simplegeneric==0.8.1
simplejson==3.11.1
six==1.11.0
sparkmagic==0.12.4
SPARQLWrapper==1.8.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
tqdm==4.18.0
traitlets==4.3.2
twine==1.9.1
typing==3.6.2
Unidecode==0.4.21
urllib3==1.21.1
wcwidth==0.1.7
webcolors==1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.1
pyswip-alt==0.2.3
networkx==1.11
reasoner
ontobio
jsonpath-rw==1.4.0
requests-cache==0.4.13
requests-cache==0.4.13
